+++ beautifycrack/src/net/beautifycrack/dao/NewsMapper.java
@@ -34,7 +34,7 @@ public interface NewsMapper
      * @return
      * @throws BusinessException
      */
-    News showNews(Integer id) throws BusinessException;
+    News showNews(Long id) throws BusinessException;
 
     /**
      * ²éÑ¯×ÜÊý
